fix: Enable streaming output for JSON and CSV reporting.
This uses the Flask `stream_with_context()` facility which keeps the
context and database session around even after the request has been
finished.

Note that we're returning CSV data as `text/csv` with a
`Content-Disposition: attachment` header, which is how it is supposed
to be done, rather than as `application/force-download`, which is a
nasty hack that doesn't belong in family-friendly open-source
code. Think of the children!1!!!1!