test: Patch the CSV test results to fit reality.
Due to changes in dictionary processing, the order in which the
columns of the CSV output are returned differs between Python 2 and
Python 3, so we rearrange the expected test results to match.

Doing direct textual comparisons on CSV data like this is something
that should nauseate any self-respecting programmer but repairing it
properly would be a considerable amount of hassle that we don't have
time for just now, so, like oh so very often, we go for the quick and
dirty fix instead.