fix: Fix `/user/enroll` reporting.
The `test_reporting.py::TestReportingController::test_selfservice`
test tries to ascertain that enrolling a new token leaves an entry in
the `Reporting` table, but this didn't work (after the `enroll`
operation, the `Reporting` table was empty).

We can trace this back to lines 363-8 in `userservice.py`; this is the
`__after__()` handler that, further down, takes care of calling the
`token_reporting()` function from `lib/reporting.py` with the realm
given in `c.audit['realm']`. Unfortunately this realm is duly
overwritten with an empty string in line 368 when the `__after__()`
handler notices that `request_context['authUser']` is not
defined. This sucks because the realm in `c.audit['realm']` is
actually perfectly OK before it is stomped on by the `__after__()`
handler, and this is a problem because with this wrong empty realm
name, `token_reporting()` doesn't find the policy which would cause
the enrollment operation to be entered into the `Reporting` table. So,
empty table and the test fails.

The root cause of the issue is that, for the `/user/enroll` operation,
`request_context['AuthUser']` isn't set up properly to begin with. If
we do that explicitly in the `__before__()` handler where
`self.authUser` is determined, then Bob's your uncle.

_we reuse the request_context['AuthUser'] which is normaly filled by
getUserFromRequest(). As we identify the user in the userservice from
the cookies this is semanticaly on the same level_