feat: add notification for autoenrollment
of email token - this fixes the restriction, that during
autoenrollment the pin of the enrolled token was set to
the given pin which is the users password. instead a
random pin is generated that will be part of the first
notification message. the second submitted notification
will contain the otp value.