fix: Token confusion that resulted in Selenium test timeout.
Make sure that `mozart` tries to reset the mOTP PIN on his own token
rather than that of `beethoven`, which happens to be left in `token`
after the everybody-sets-their-PIN loop in line 211-4 above. This
probably used to work in Python 2 because the right token ended up in
`token` due to a fortunate accident in dictionary ordering -- but in
Python 3.7, the ordering of items in `user_token_dict` is preserved,
so the last one through the loop is `beethoven`.