fix: Policy deletion in Selenium now waits for the UI to update
Previously, the routine would click on the policy line and then
click delete without waiting for the policy name field to update.
There is a small window of time between the click on the policy
and the UI update, and if we click at that point the UI sends an
empty policy deletion request.

We fix this by waiting for the policy name input box to be filled
after the click.