refactor: Remove the `linotp/config` directory.
The `linotp/config` directory is, for the most part, a holdover from
the Pylons age. Most of the functions defined there are now stubs that
refer to Flask stuff, and we move their content to where the functions
were called (insofar as that hasn't been done already). The
`config/routing.py` file is no longer being used anywhere and can be
removed outright, as is the `config/environment.py` file.

We move `config/defaults.py` up one level to the `linotp` directory
where it can stay for the time being.