feat: allow token test to support challenge response mode.
A first request with a token serial is used to announce a new token test.
The server decides, depending on the token type, what the next step is. If
the token requires a challenge transaction, the request responds with a
transactionid. If available, this transactionid must be sent along with the
otp in a second request.

Refers to LINSELF-15