refactor: identify the verify action upfront
to make the code better understandable. In addition
all handling of transaction to challenge identification
was moved to the first place of transactioid test.

in addition this code contains a small fix for the response
preparation which is different as the validation handlers directly
return the boolean value instead of the token counter