fix: Avoid problems with `/system/getPolicy/policy.cfg`.
The `/system/getPolicy` endpoint supports optionally passing a file
name at the end of the URL. This causes the policy to be sent as a
`text/plain` resource attachment with the given file name as a
suggestion for local storage. If the file name is omitted from the
URL, the policy is returned as an `application/json` resource
instead. In the web-based management UI, `/system/getPolicy` is used
to fill in the policy detail dialog on the “Policies” tab, while
`/system/getPolicy/policy.cfg` is used to power the “Export Policies”
button.

The problem with this setup is that elsewhere in the code the final
part of a URL path is used as a “method name” for URL dispatch and
policy checks. If a request is made to `/system/getPolicy/policy.cfg`,
LinOTP tries to do a policy check for a non-existing `policy.cfg`
method, which fails.

We mitigate this problem by removing `/policy.cfg` as an optional
third part of the URL path and introducing an `export=true` query
parameter instead. This means that `/system/getPolicy` is used as the
API endpoint in every case, and the content type of the resulting
resource depends on whether this query parameter is passed.

An alternative implementation would exploit the `Accept:` header to
figure out the appropriate content type for the response, but this
would probably be too invasive as an after-the-fact change.