fix: in admin and userservice controller for required parameters
which will be required within the checkPolicyPre. This is either
for the assignment the check that a serial number must be part of
the parameters or the 'type' parameter which is a 'must' for the
enrollment.
This is required as the checkPolicyPre is a generic method, which
only relies on the param parameter, which contains arbitrary key
value pairs.