feature: removal of the local cert file handling
which is not recomended due to the problems with the temporary
cert file, eg. during the transition from testing to the normal
processing.
In general we follow the assumption that the ldap resolver should
rely on the system certificate store, which is the common place to
deal as well with other tls requests like https.