feature: remove the ocra backend interfaces
which are the ocra controller that was used for the old
ocratoken as a dedicated interface to enroll and check the
transaction as it was the first challenge response token
with stored challenge data.
and the userservice / selfservice interfaces to enroll and
activate the ocra token.
accordingly the acl policies for the ocra controller and
the userservice policies are removed.