fix: for preserving  selfservice auth secret
which s used to salten the hmac that is used for the
userservice cookie based authentication.

the described problem occures in an HA scenario, where
when using the slefservice, the config entry is created
if it does not exist. It seems to be a race condition
where the entry is written twice from each HA partner
which results in an conflict as the key would be created
twice, which violates the unique constrain on the config
key names.

the solution is to not use the config table persistancy
as the seecret could be created new on every server start.

the solution corresponds to the current handling of the
userservice cookie store, which is bound to the server
runtime.

refers to LINOTP-1328