refactor: Replace explicit test loop with `@pytest.mark.parametrize`.
We don't like test functions that run loads of tests in a loop,
because if one of them fails it isn't easy to figure out which one it
was and the remaining ones aren't run. It's a lot better to use pytest
parametrisation for this because then all the `assert` statements
happen in individual tests.