feature: Ensure `SQLIdResolver.check_password` catches `MissingBackendError`.
If there is no backend library installed for the `bcrypt` support in
`passlib`, `passlib` raises `MissingBackendError`. So far this has not
been made explicit in the error log. The “Failed to check user
password” message there didn't aid in pinpointing the actual issue,
and an obscure message stating `bcrypt 'builtin' backend not enabled
via $PASSLIB_BUILTIN_BCRYPT`, which would appear at the point where
the `bcrypt` functionality was first accessed, isn't entirely helpful
either.

We change the `SQLIdResolver.check_password()` function to log an
explicit error if the `passlib` complains about a missing backend (not
just for `bcrypt`, either) when a password is being checked. This
should cause a more helpful message to show up in a place where it
catches an administrator's eye.