fix: rendering problems in selfservice for IE11
IE11 had problems with rendering our flexbox layout in selfservice.
It is used to stretch the application to the full height of the screen.
IE11 won't be able to fully support our design, but it now falls back
to a good looking solution, just not as good looking as other browsers
can do it. 🥷 css skills would be neccessary to make IE11 work...

Closes IE11 😉
Fixes LINOTP-1313