feature: Add machinery to copy packages from buster to buster-golden.
We want to do this when a commit has a `release/…` tag (but we also
allow ourselves to trigger the operation manually).

The `buster-golden` distribution on the internal Debian-style
repository server is used to hold “golden” packages which have passed
QA and are ready to be released to the official mirror. This approach
uses `reprepro` to copy packages from the regular `buster`
distribution to `buster-golden` when it detects that a package that
was just uploaded is a “golden” package. Actually moving packages from
`buster-golden` on the internal repository server to the official
mirror is beyond the scope of this changeset.