feature: prepare token base to handle token type challenge validity
LinOTP support since a while token type specific challenge validity
times by config entries like 'EmailChallengValidityTime' but for
setting these entries there had been no gui support.
As many customers ran in the problem of setting these value for SMS
or EMail token in combination with radius, it is a long standing feature
to be able to set these values via manage ui.

In LinOTP therefor the mako token specific templates allow to
define theres setting easily. But the implementation has the constrain
that the config values are prefixed by the token type in lower case.

Thus we have to extend the determination of the challenge validity time
to support historically the caps config and the lower case prefix
config value like 'emailChallengeValidityTime'.