fix: remove sms local class method to get the challenge validity time
which will be handled in the token base class for all tokens.
the specific query for the provider timeout to derive the
challenge validity time is outdated, commimg back from a time
where only one provider existed. thus the code could be dropped
and be replaced with the common method from the base class.