fix: monitoring test with ambigous user 'root'
as there is a root user in both resolvers. the problem comes
with the assignement of the token to the user root, which now
is ambigous - one of ther resolver got the match which has been
a different one when the test was written.
to fix this the root user was replaced with the user 'susi'
which is unique wrt to the resolver.
in the standard manage ui the token assignement does not
have this problem as in addition to the realm and the user name
the resolver of the user is provided