feature: Add skeleton for API documentation.
This adds an `api-doc` directory with a few configuration files for
Sphinx. Most of the API documentation is generated dynamically from
the source code by means of the `sphinx-apidoc` command, and written
to the `source` directory.

There is a rudimentary Sphinx theme based on the `default` theme. This
has been mostly adopted from the earlier API doc support in
`linotp-doc`.