refactor: Replace homegrown I18N tooling with Flask-Babel.
This gets rid of large swathes of code implementing functionality that
is more conveniently provided by Flask-Babel. In the spirit of
sticking to industry-standard best practices, we're using functions
like `gettext()` directly instead of inventing our own names to
obfuscate things.

We're passing `lang` into the `render_template()` function instead of
figuring it out in the template because that type of logic doesn't
belong in templates, and also DRY.

The message catalog root directory and name are configurable in the
Flask settings because Flask-Babel likes it that way. We put them in
the `settings.py` file so people could theoretically tweak them but
this is definitely a “hands off unless you know what you're doing”
sort of thing. Note that `BABEL_TRANSLATION_DIRECTORIES` is in fact
a search path, so this opens the door to additional out-of-tree
translations, even though that might need a little work in
`linotp_utils.js`.

Also note that the JSON message catalogs are now loaded from
`/static/i18n/` instead of `/i18n/`, which is correct because all
static files for the UI are served from below `/static/`. We really
don't want to use Flask to serve those at all unless perhaps in a
development setup, because serving dynamic and static resources from
the same process is a rookie mistake, but that's a fight for another
day.