refactor: Fix I18N in Mako templates.
The `lang` value is now passed into the Flask-Mako `render_template()`
function inside our applicaton so we don't need to figure it out
explicitly several times over in various templates. That type of logic
has no place in templates, anyway.

Nobody seems to know what `window.ALL_CURRENT_LANGUAGE` is supposed to
be good for, so we're getting rid of that as well as the `allang`
value.

We don't need the `isinstance()` guard because Flask-Babel guarantees
us that `lang` is a string.