feature: Special pathname handling for `*_DIR` and `*_FILE` settings.
Configuration settings that end in `_DIR` or `_FILE` are assumed to
contain path names. If the value of such a setting starts with a
slash, it is used as is; otherwise we prepend the value of the
`ROOT_DIR` setting to it. (This is basically what Apache is doing,
too.)

By default, `ROOT_DIR` is the directory that contains the `linotp.cfg`
file; if there is no such file, the parent of the directory containing
the `settings.py` file is used instead (which turns out to be the root
directory of the LinOTP software distribution).

As a special case, the value of the `BABEL_TRANSLATION_DIRECTORIES` is
a string containing a list of directory names separated by semicolons.
Individual directories in this list also have the value of `ROOT_DIR`
prepended unless they are absolute.

Note that the configuration settings `AUDIT_PUBLIC_KEYFILE` and
`AUDIT_PRIVATE_KEYFILE` are now called `AUDIT_PUBLIC_KEY_FILE` and
`AUDIT_PRIVATE_KEY_FILE`, to match the `*_FILE` rule.