feature: Support list of configuration files.
With this change, LinOTP consults a list of configuration files which
are read in order (left to right); later configuration items (lower in
the same file or in a file that is read afterwards) overwrite earlier
ones. If a file named in the list cannot be read, this elicits a
warning on the console but is otherwise not a show-stopping problem in
itself.

The list is stored in the `LINOTP_CFG` environment variable. If that
variable is not set, LinOTP looks for a `linotp.cfg` file in the Flask
`app.root_path`, which for us amounts to `…/linotpd/src/linotp` within
the repository.