refactor: Introduce `ConfigSchema` for configuration items.
Instead of using a Python class for the hard-coded configuration
defaults (with explanatory comments and ad-hoc typing) we're using a
special `ConfigSchema` to describe the set of valid configuration
settings, with associated types, checking and conversion functions,
the default values, and a help text. This will make it possible to
convert configuration values from strings (e.g., in environment
variables) to their proper values, and determine which configuration
settings have been changed from their default values, among other
things. Other than that there is no difference to the previous setup;
in particular, the configuration files are still Python code.

This commit replaces the traditional `Config` class with an approach
based on the `ConfigSchema` and adds validation against the schema to
the `LinOTPConfig.__setitem__()` method. For the time being we're only
treating the configuration items that had hard-coded defaults in the
old `Config` class. Configuration items which are not in the schema
still work, they just can't make use of the benefits that the schema
affords registered items.