feature: Add support for `CUSTOM_TEMPLATES_DIR` (ex `custom_templates`).
This hadn't made it across from LinOTP 2.x earlier. It is slightly
dicey in LinOTP 3 because Flask doesn't really support search paths
for templates (you get to specify one template directory for the app
and then additional ones for blueprints). It turns out, however, that
Flask-Mako doesn't mind if `app.template_folder` is a tuple or list.

Note that the new configuration setting is called
`CUSTOM_TEMPLATES_DIR` so our nifty prepend-`ROOT_DIR` mechanism will
work. Also note that we've tweaked `LinOTPConfig.__setitem__` such
that a `*_DIR` or `*_FILE` setting with a default value that is empty
or `None` won't croak.