feature: Add support for `CACHE_DIR` (ex `cache_dir`).
This is the directory used for various resolver caches. It should
probably not be confused with `BEAKER_CACHE_DIR`; while we're at it we
rename the default `BEAKER_CACHE_DIR` to `beaker_cache` so there is no
possibility for conflict between the contents of `CACHE_DIR` and
`BEAKER_CACHE_DIR`. We also remove the hard-coded defaults from the
Beaker initialisation because we don't need them anymore -- all
hard-coded defaults are in `settings.py`.