feature: Change `ROOT_DIR` default and add validation.
This removes the distasteful `/ROOT_DIR_UNSET` default value for
`ROOT_DIR`, which is unnecessary since the actual “default value” for
`ROOT_DIR` is set to `app.root_path` in the
`settings.Config.init_app()` method, so `/ROOT_DIR_UNSET` never
actually goes anywhere.

While we're at it, we add a validation function which ensures that the
value of `ROOT_DIR` is, in fact, an absolute path name (i.e., a
non-empty string that starts with a slash).