chore: Enable DNS resolution between services in CI
Gitlab runner currently uses legacy container links:
 https://docs.gitlab.com/runner/executors/docker.html#legacy-container-links

However in this mode, the 'alias' directive only works for the runner
itself. But we need to spin up a linotp service that depends on
another service (the database).

So we switch to network per build mode
  https://docs.gitlab.com/runner/executors/docker.html#network-per-build

Note that this requires gitlab-runner version >= 12.9