refactor: Move configuration machinery out of `linotp.app.create_app()`.
We move everything that is to do with configuration of the Flask
app (establishing the hard-coded defaults, reading configuration
files, checking environment variables for further configuration) from
the `create_app()` function into a separate `configure_app()`
function.

This makes the configuration machinery easier to test, because we can
invoke `configure_app()` to check whether it does the Right Thing™
without having to go through the complete app-creation song and dance
routine.