refactor: Rewrite `LDAPIdResolver.IdResolver.connect()`.
We're trying to make the LDAP server connection code less convoluted
and more logical, and we're also adding some comments for clarity.  In
addition, we try to be more clear in log messages to tell people what
is really going on.

Note that, for the time being, we keep the strange behaviour where
LinOTP tries to do STARTTLS on plain LDAP connections even if the
administrator explicitly asks it not to (with no certificate checking)
and forces certificate checking on LDAP with explicitly solicited
STARTTLS even if the configuration says not to do it. If that ever
changes we need to remove some clearly-marked code.

Also note that the trust store to be used for root-level CA
certificates can now be specified globally via a configuration setting
in `settings.py`. (This probably doesn't need to be more granular
because you can always make your own trust store that has all the
root-level CA certificates you need.)