feature: Add `linotp ldap-test --all-cases`.
This does an exhaustive test of all combinations of LDAP,
LDAP+STARTTLS, and LDAPs with certificate checking enabled or disabled
and a root-level CA certificate available or not. We're bending over
backwards to present the results in a convenient table that is, as
much as possible, unsullied by logging output or tracebacks (hence the
`silent` treatment for `resolver_request()` a few commits back).

Debugging messages from the LDAP resolver are enabled iff the general
logging level (which can be set via the `LINOTP_LOGGING_LEVEL`
environment variable when calling `linotp ldap-test`) equals
`DEBUG`. This lets you chase down problems more easily if any of the
cases don't behave as they should, but you can of course always diddle
the `-e`, `-c`, … options as required outside of `--all-cases` to dig
into any issues.