test: Add integration test for LDAP connection methods.
This test leverages the `linotp ldap-test … --all-cases` command to
check that all the different ways to connect to an LDAP server yield
the expected results. In particular, this includes making sure that
connection attempts fail if we insist on checking server certificates
that turn out to be invalid, which is impossible to do via Selenium.