feature: Add the `linotp config show` command.
This adds the `linotp config show` command, together with some basic
infrastructure to dispatch to the `linotp config` command group. (It
turns out that `flask.commands` entry points are happy to dispatch to
command groups, which is nice.)

Note that we get the Flask application context for free because
`config_cmds` is an `AppGroup`, rather than a `click.group`.