chore: Remove debian packaging override for pycryptodomex
Debian has provided the Python Pycryptodomex package since Buster
in the package python3-pycryptodome, so we no longer need to include
this as a separately packaged dependency.

By removing the pydist-overrides file, we revert back to the standard
debian package resolution system, which will automatically pull
python3-pycryptodome in for us.