feature: Enable WSGI request profiling if `PROFILE` is `True`.
We enable the Werkzeug request profiler if the `PROFILE` configuration
item is set to `True`. This is probably best done on an ad-hoc basis
for the development server by saying `LINOTP_PROFILE=true linotp run`.

Things are complicated slightly because depending on your Werkzeug
version the profiler can be in different modules.