refactor: Remove audit data from `request_context`.
The audit log is initialised by instantiating the appropriate
audit-log access class (either `SQLAudit` or nothing) and putting the
resulting object into `request_context['config']['audit']`. While a
request is processed, interesting tid-bits for the audit log are put
into the `request_context['audit']` dictionary (usually disguised as
`c.audit`), which is written to the audit log via the
`request_context['config']['audit']` object when the request is
done. If this isn't confusing enough, the
`request_context['config']['audit']` object is also put into
`request_context['Audit']` for no discernible reason.

We clean this up by (a) storing the audit-log access object as
`app.audit_obj`, and (b) using the Flask request global object (`g`)
to store data to be written to the audit log (in the `g.audit`
dictionary). This highlights the fact that the audit log persists
while the server is running, while audit data is ephemeral and is
removed after a request has been dealt with.

The advantage of this approach is that `g` is fairly well-understood
in Flask circles, while nobody can say exactly what `request_context`
in LinOTP is and how and why specific stuff ends up in it. (It is a
holdover from the Pylons age and by now basically consists of a bunch
of abstraction layers on top of, you guessed it, `g.request_context`,
but the more stuff we can pull out of it and put into a simpler place,
the better.)

Of course we need to change the controllers to follow suit, and will
do that anon.