refactor: Move `setup_audit()` out of the code path for every request.
This is what this is all about: We remove the call to `setup_audit()`,
which is horrendously expensive, out of the code path for every
incoming request to where it is called only once when the server is
initialised. This makes a considerable difference to the efficiency of
the server.