test: Move configuration within a temporary directory
The goal here is to not require any readable/writable configuration
files anywhere outside of the source tree we are testing from. This
is to make sure that any tests are easily reproducible.

By no longer pointing to an existing directory, we have the problem
that we need some encryption keys for the audit and database. We
move this setup to a per session fixture to make sure it is only
done once. Another alternative to consider would be to create some
keys in the fixtures directory.

* Prevent usage of LINOTP_CFG in the environment so we are independent
  of the current packaging
* Point ROOT_DIR and LOGFILE_DIR to a temporary directory
* Point the audit and enc key file to a per-session directory