fix: Make `ExtFlaskConfig.get()` go through `….__getitem__()`.
Our special relative-pathname hack for configuration items whose names
end in `…_FILE` or `…_DIR` is implemented in the
`ExtFlaskConfig.__getitem__()` method, but `ExtFlaskConfig.get()`
doesn't use `__getitem__()` internally. (This seems counterintuitive
but the probable explanation is that `__getitem__()` is there to make
indexing with brackets work no matter what actually happens with the
data.)

We're redefining `ExtFlaskConfig.get()` to actually use bracket
notation to get at the value, so our special magic will apply. We need
to take care to return the optional default if the key doesn't
actually exist.

Note the flaming hoop we need to jump through in `__getitem__()` to
make `r = config['ROOT_DIR']` work without generating an infinite
regression.