refactor: Replace Flask-Mako with homegrown interface code.
This change replaces the use of Flask-Mako in LinOTP with simpler,
less general code that interfaces with Mako directly. This is not
because we don't like Flask-Mako; the main reasons are that in the
absence of an explicit license grant, the licensing situation of
Flask-Mako is unclear, and that the project doesn't appear to be
actively maintained.