feature: Add `linotp.cli.get_backup_filename()` function.
Given a file name, this returns a timestamped backup file name. This
can be used to save an older version of a file or to create a
timestamped file for a backup copy, among other things.

This commit anticipates LINOTP-1424, but we're not changing the
existing code to use it, because that would be too blatant. Nothing in
LINOTP-1403, however, prevents us from implementing this for *new*
code :-)