test: [docker] Move softhsm setup to new linotp-softhsm image
Deployments of linotp using a real Hardware Security Module will
need to provide the relevant libraries and configuration to use
the HSM from within Docker. Thus they will need to derive from
the linotp image.

In order to avoid confusion, we remove usage of the SoftHSM from
the main linotp images, since this implementation is not suitable
for real production use. This is becuase multiple requests cannot
be carried out in parallel.

So given the above, we refactor the softhsm configuration to define a
derived Docker image that contains a sample softhsm configuration. This
is only for testing, so the configuration is mostly hardcoded.

Closes LINOTP-1419