feature: Clean up `linotp init enc-key` command.
We extract the verification and file-backup steps of the command into
separate functions so we can reuse that functionality elsewhere. This
also simplifies the implementation of the `linotp init enc-key`
command itself.

In addition, we add the `--dump` option, which outputs a textual
representation of the enc-key that is reasonably straightforward to
re-enter by hand. The `--keys` option allows a user to re-enter a key
from the representation written by `--dump`, which contains CRC32
checksums and a SHA1 hash to make it easier to find typoes.