feature: Add `linotp init audit-keys` command.
This replaces the functionality of the `AuditBase.createKeys()` method.

Note that when the audit log is initialised, this calls `sys.exit(11)`
if the private *or* public part of the audit log key pair doesn't
exist. (The previous version of the code would complain only if *both*
parts were missing, which wasn't strict enough. If only one part is
missing, the code would fail later in any case.)