feature: Change default for `ADMIN_{USERNAME,PASSWORD}` to empty string.
This used to be `None`, but the empty string is easier to set from an
environment variable.

The code used `app.config.get(…)` to obtain the values for the
settings, which defaults to `None`, but we don't want to scatter
default values for configuration settings all over the code base
anymore, so now it's `app.config[…]`, where the default comes from
`settings.py`.