test: Tweak testing infrastructure to work with `db_setup()` split.
We want to make sure that the tests still work even in the absence of
explicit calls to `linotp init database`. This means that we have to
pretend that `linotp init database` has been called when the test
environment is initialised.

The testing setup for the `init` command comes with its own `app`
fixture that doesn't initialise the database so that still works.