test: Fix `Dockerfile` for end-to-end tests.
We can't call `linotp init all` without `--force` because the enc-key
and audit keys are already created in the Debian package
`postinst`. For the time being, we just use `linotp init databases` to
initialise the database structure.

We use this opportunity to also fix some of the environment-based
configuration items that were misspelled.