feature: Omit unreadable-config warning if name starts with dash.
When reading files mentioned in the `LINOTP_CFG`, a warning will be
output if the file in question couldn't be read (e.g., because it
doesn't exist or the reader has insufficient permissions). This
discourages distributions from including files in `LINOTP_CFG` on the
off-chance because the superfluous warnings could be confusing.

This change omits the warning if the name of a file in `LINOTP_CFG`
starts with a `-` (by analogy to makefiles and systemd unit
files). The dash is not considered to be part of the filename
itself. IOW, given

    LINOTP_CFG=/foo/linotp.cfg:-/bar/linotp.cfg:/baz/linotp.cfg

the warning will be output if there are issues with `/foo/linotp.cfg`
or `/baz/linotp.cfg` but will be omitted for `/bar/linotp.cfg`.

Note that, from a backwards-compatibility POV, this will make it
impossible to have configuration files that are (a) specified using a
relative path name, (b) whose relative path name starts with a dash,
and (c) where we want the warning shown. The chances of this occurring
in actual practice are infinitesimal.