fix: Fix another C&E issue when `SQLALCHEMY_DATABASE_URI` isn't known.
We tried to catch the case where the database tables aren't
initialised (and make a special exception for `linotp init …`), but it
turns out that everything goes south when `SQLALCHEMY_DATABASE_URI`
isn't initialised properly, too.

Our quick fix is to bury our head in the sand as far as the database
is concerned when executing `linotp init …` commands,none of which
actually *use* the database, apart from (duh!) `linotp init database`,
but that initialises the database engine explicitly when it comes to it.