refactor: Remove `BEAKER_CACHE_DIR` configuration setting.
We use `CACHE_DIR/beaker` instead, in order to cut down on
configuration settings.

At the same time we move the resolver cache, which would otherwise end
up in `CACHE_DIR`, to `CACHE_DIR/resolvers`. This is to safeguard
against the inevitable day when someone will come up with the glorious
idea of calling a new resolver `beaker` (perhaps in honour of Dr
Bunsen Honeydew's lab assistant) and decide to give it a cache
directory.