fix: Make sure config items from `create_app()` are set properly.
The second parameter to `create_app()` can be a dictionary of
configuration settings. This used to be stuffed into the app config by
means of `dict.update()`, which meant that our carefully crafted
`ExtFlaskConfig.__setitem__()` method would not be used, and in
particular, file and directory names would not get the special
treatment they're due.

We're giving `ExtFlaskConfig` objects their own `update()` method
which ensures that `ExtFlaskConfig.__setitem__()` is called for every
item.