fix: Guard against `pathlib.Path` objects in configuration dicts.
Pytest users like to take advantage of fixtures like `tmp_path`, which
are instances of object-based representations of path names, rather
than strings. This can lead to problems because by and large, the
configuration subsystem assumes that incoming configuration items
consist of strings, and says things like `value[0] == '/'`. To avoid
problems of this type (no pun intended), we convert the values of
settings that refer to directories and files to strings, on the
off-chance.