feature: Introduce `DATA_DIR` configuration setting.
This is intended for data that arises during LinOTP runs but isn't
expected to hang around. The directory can be put below `/run`.

The commit also includes the changes to the Debian packaging to ensure
that `DATA_DIR` ends up in `/run/linotp/data`. These consist of a
`debian/linotp.tmpfile` file which – thanks to the magical properties
of `dh` – will end up in `/usr/lib/tmpfiles.d/linotp.conf`, a
dependency on `systemd` in `debian/control` so we know that
`systemd-tmpfiles` will be available, and a stanza in
`/usr/share/linotp/linotp.conf` that moves `DATA_DIR` to
`/run/linotp/data`. This works because once we have a `/run/linotp`
file that belongs to the `linotp` user, which `systemd-tmpfiles` will
make on our behalf when the system is booted, based on the content of
`/usr/lib/tmpfiles.d/linotp.conf`, we can create subdirectories at our
convenience. The `postinst` script also calls `systemd-tmpfiles`
automagically so that the directory is available immediately after the
package has been installed.