feature: Check on startup that important directories exist.
In order to avoid nasty surprises later, we check that important
directories exist early during the configuration phase. This includes
ROOT_DIR, DATA_DIR, CACHE_DIR, and LOGFILE_DIR. It is the
administrator's responsibility to ensure that these directories are
present; the Debian package will install them but other users will
have to take care of this themselves.

The directory check is omitted when we're executing `linotp config`
commands. This makes it possible to do things like

    $ sudo mkdir -p $(linotp config show --values DATA_DIR CACHE_DIR)

to create the directories in question if one doesn't know by heart
where they should go.