feature: Croak if schema isn't installed, except when doing `linotp init database`.
We want LinOTP to crash with a fatal error if the main database schema
hasn't been installed properly, except when we're initialising the
database.

The problem here is that we need to install the database schema all
the time during testing, so we need to pretend we're doing `linotp
init database` when the testing harness is running `create_app()`,
even if we actually want to do something else. This is not a problem
because when we're running tests for CLI commands, the app is created
first and the command line is faked later.