refactor: Rip out the homegrown DB stuff and insert `flask-sqlalchemy` stuff.
The main difference is that most of the database stuff hangs off the
global Flask-SQLAlchemy `db` object instead of a bunch of global
variables that we have defined ourselves.

Most of the remainder of this MR will deal with replacing references
to the homegrown stuff elsewhere in the code by references to `db`.

Note that for the time being we're disabling the database migration
code; this will come back in due course.