refactor: Fix `Session` references in controllers.
We replace `Session` by `db.session` and get rid of extraneous
`Session.close()` requests, which Flask-SQLAlchemy will handle for
us. (Having the `Session.close()` calls in there may actually be
detrimental, so it's better to get rid of them.)

Note that we could have said `Session = db.session` in a strategic
place and left most of this code intact, but it's better to rip the
band-aid off. A well-known saying in computer science claims that
“every problem can be solved by adding another layer of abstraction”,
but we want to go back to the basics – LinOTP is confusing enough
already and we don't need to add to the spaghetti bowl. At least with
`db.session` we know where we stand, and our code looks like a Flask
application.