test: [test-upgrade] add tool for install and upgrade testing
This is a tool which I have been using to test various package
install and upgrade scenarios.

It needs to be run on a system with access to the apt database
(sudo is ok) and pytest for the Selenium tests. Configuration
to match the local system can be placed in a '.env' file.

The script is not yet integrated into the CI pipelines. It needs
more work before that can be done, but it is useable for local
testing already.