feature: Read config file list from file if `LINOTP_CFG` doesn't exist.
Normally the list of configuration files to be read is taken from the
`LINOTP_CFG` environment variable, with a hard-coded default of
`linotp.cfg` in the “app root” if the variable isn't defined. This
makes it difficult to establish a sensible default for command-line
usage in Linux distributions. (WSGI usage is covered because the
environment variable can be set in the WSGI file.)

This patch adds code to read the default value for the list of
configuration files from the `linotp-cfg-default` file in the “app
root” directory if `LINOTP_CFG` is not defined but that file
exists. If the file doesn't exist, LinOTP still falls back on
`linotp.cfg` in the absence of `LINOTP_CFG`.